requirements.txt
pymupdf
pillow
pyinstaller
openpyxl